requirements.txt
amply==0.1.4
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
comtypes==1.1.10
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-mailer==2.1
docutils==0.17.1
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==2.10
kiwisolver==1.3.1
kombu==5.1.0
lockfile==0.12.2
matplotlib==3.4.2
mpmath==1.2.1
numpy==1.20.3
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
PuLP==2.4
pycryptodome==3.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==1.5
pyttsx3==2.90
pytz==2021.1
reportlab==3.5.68
requests==2.25.1
scipy==1.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
sympy==1.8
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wikipedia==1.4.0
xhtml2pdf==0.2.5
